+++ src/sisl/physics/hamiltonian.py
@@ -7,7 +7,7 @@ import sisl._array as _a
 from sisl._internal import set_module
 
 from .distribution import get_distribution
-from .electron import EigenstateElectron, EigenvalueElectron, spin_squared
+from .electron import EigenstateElectron, EigenvalueElectron
 from .sparse import SparseOrbitalBZSpin
 
 __all__ = ["Hamiltonian"]